+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+ * use poetry name if available
+
 ## [0.4.1] - 2019-08-06
 
 ### Added